injozi backend project/myproject/requirements.txt

bcrypt==4.0.1
blinker==1.6.2     
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Jinja2==3.1.2
jwt==1.3.1
psycopg2-binary==2.9.7
SQLAlchemy==2.0.19
Werkzeug==2.3.6